agentic-ai-engine/pyproject.toml
[tool.poetry.dependencies]
python = "^3.11"
fastapi = "^0.112.0"
uvicorn = {extras = ["standard"], version = "^0.30.0"}
gunicorn = "^22.0.0"
slowapi = "^0.1.9"

pydantic = "^2.8.0"
pydantic-settings = "^2.4.0"
httpx = "^0.27.0"
orjson = "^3.10.0"
python-dotenv = "^1.0.1"
structlog = "^24.2.0"
